requirements.txt
python-dotenv==1.2.2
scikit-learn>=1.4.0
openai==2.24.0
playwright==1.58.0
pillow==12.1.1
SQLAlchemy==2.0.48
requests==2.32.5
PyMuPDF==1.27.1

# Optional extras only if that OCR pipeline is enabled in dte_loader:
# easyocr
# opencv-python
# ultralytics
# transformers
